collaborative-planner/backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-socketio==5.3.4
flask-cors==4.0.0
flask-login==0.6.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
email-validator==2.0.0
gunicorn==20.1.0
